move class ServerProxy to package "player" and added drag-sort-list-view library
