+++ app/src/main/java/com/irateam/vkplayer/serveces/AudioService.java
@@ -1,0 +1,4 @@
+package com.irateam.vkplayer.serveces;
+
+public class AudioService {
+}
